Remove xmlPreprocessor, move it to config reducer #4886
