+++ ui/app/src/state/reducers/dashboard.ts
@@ -17,7 +17,7 @@
 import GlobalState from '../../models/GlobalState';
 import { createReducer } from '@reduxjs/toolkit';
 import { INIT_DASHBOARD_CONFIG } from '../actions/dashboard';
-import { deserialize, fromString, xmlPreprocessor } from '../../utils/xml';
+import { deserialize, fromString } from '../../utils/xml';
 import { applyDeserializedXMLTransforms } from '../../utils/object';
 
 const reducer = createReducer<GlobalState['dashboard']>(null, {
@@ -25,12 +25,10 @@ const reducer = createReducer<GlobalState['dashboard']>(null, {
     let dashboardConfig = null;
     const arrays = ['widgets', 'roles', 'excludes', 'devices', 'values', 'siteCardMenuLinks', 'tools'];
     const renameTable = { permittedRoles: 'roles' };
-    const configDOM = xmlPreprocessor(fromString(payload.configXml), payload.references);
+    const configDOM = fromString(payload.configXml);
     const dashboard = configDOM.querySelector('[id="craftercms.components.Dashboard"] > configuration');
 
     if (dashboard) {
-      // TODO: keys should go in xmlPreprocessor
-      dashboard.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
       dashboardConfig = applyDeserializedXMLTransforms(deserialize(dashboard), {
         arrays,
         renameTable
