+++ ui/app/src/state/reducers/launcher.ts
@@ -17,7 +17,7 @@
 import GlobalState from '../../models/GlobalState';
 import { createReducer } from '@reduxjs/toolkit';
 import { INIT_LAUNCHER_CONFIG } from '../actions/launcher';
-import { deserialize, fromString, xmlPreprocessor } from '../../utils/xml';
+import { deserialize, fromString } from '../../utils/xml';
 import { applyDeserializedXMLTransforms } from '../../utils/object';
 
 const reducer = createReducer<GlobalState['launcher']>(null, {
@@ -25,11 +25,10 @@ const reducer = createReducer<GlobalState['launcher']>(null, {
     let launcherConfig = null;
     const arrays = ['widgets', 'roles', 'excludes', 'devices', 'values', 'siteCardMenuLinks', 'tools'];
     const renameTable = { permittedRoles: 'roles' };
-    const configDOM = xmlPreprocessor(fromString(payload.configXml), payload.references);
+    const configDOM = fromString(payload.configXml);
 
     const launcher = configDOM.querySelector('[id="craftercms.components.Launcher"] > configuration');
     if (launcher) {
-      launcher.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
       launcherConfig = applyDeserializedXMLTransforms(deserialize(launcher), {
         arrays,
         renameTable
